SiLU関数の指数関数呼び出しを修正し、精度を向上させるためにexpをexpfに変更

+++ act.c
@@ -21,7 +21,7 @@ Tensor* RELU6(Tensor* input_tensor){
 
 Tensor* SiLU(Tensor* input_tensor){
     for(int i=0; i < input_tensor->H * input_tensor->W * input_tensor->C; i++){
-    input_tensor->data[i] =  input_tensor->data[i] / (1 + exp(- input_tensor->data[i])); 
+    input_tensor->data[i] =  input_tensor->data[i] / (1 + expf(- input_tensor->data[i])); 
   }
   return input_tensor;
 }
